Fixes corrections from pull request #65
Removes the instance method attribute from annotation.
Ajusts usage in the example file.
