+++ middleware/src/main/java/br/ufrn/dimap/middleware/lifecycle/Static.java
@@ -6,12 +6,12 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Static Annotation, to be used on statictest objects that can be accessed by any client.
+ * Static Annotation, to be used on static objects that can be accessed by any client.
  * <p>
- * Objects annotated with it are initializated at server start up and it's attributes cannot be
+ * Objects annotated with it are initialized at server start up and it's attributes cannot be
  * changed by clients. These objects are also destructed on server shut down.
  *
- * @author Gilberto Soares
+ * @author Gilberto Soares, Wilson Farias
  * @version 1.0
  */
 @Target(ElementType.TYPE)
@@ -21,11 +21,4 @@ public @interface Static {
      * Defines the load priority for the annotated class. Higher numbers are loaded first.
      */
     int priority() default 10;
-
-    /**
-     * Defines the initialization method for the annotated class.
-     *
-     * @return the method name.
-     */
-    String method() default "getInstance";
 }
